Update xdrv_22_sonoff_ifan.ino
We need to change the way the relays are switched from the default way which is incorrect.

Slow (1) – relay 1 on 2 & 3 off
Medium (2) relay 1 & 2 on 3 off
Fast (3) relay 3 on 1 & 2 off

This behaviour does not give the correct speed control. To get the correct speed control we need the relays to switch as follows.

Slow (1) – relay 1 on 2 & 3 off
Medium (2) relay 2 on 1 & 3 off
Fast (3) relay 3 on 1 & 2 off

Another important feature we want to implement is that when you turn on the fan from the slow or medium speeds the hi speed relay is engaged briefly before switching to the speed that was selected. This is done to give the fan maximum power to get moving, this will save the inrush current from going through the capacitors and increase their life.

For more details see https://www.breznet.com/2021/07/31/sonoff-ifan03/